+++ src/grid/gridstorage/single-array-allocation.hpp
@@ -50,7 +50,7 @@ namespace schnek
      * @tparam T The type of data stored in the grid
      * @tparam rank The rank of the grid
      */
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     class SingleArrayInstantAllocation
     {
     public:
@@ -106,7 +106,7 @@ namespace schnek
      * @tparam T The type of data stored in the grid
      * @tparam rank The rank of the grid
      */
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     class SingleArrayInstantFortranAllocation
     {
     public:
@@ -166,7 +166,7 @@ namespace schnek
      * @tparam T The type of data stored in the grid
      * @tparam rank The rank of the grid
      */
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     class SingleArrayLazyAllocation
     {
     public:
@@ -233,20 +233,20 @@ namespace schnek
     //=============== SingleArrayInstantAllocation ====================
     //=================================================================
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     void SingleArrayInstantAllocation<T, rank>::resize(const IndexType &low_, const IndexType &high_)
     {
         this->deleteData();
         this->newData(low_, high_);
     }
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     SingleArrayInstantAllocation<T, rank>::~SingleArrayInstantAllocation()
     {
         this->deleteData();
     }
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     void SingleArrayInstantAllocation<T, rank>::deleteData()
     {
         if (data)
@@ -255,7 +255,7 @@ namespace schnek
         size = 0;
     }
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     void SingleArrayInstantAllocation<T, rank>::newData(
         const IndexType &low_,
         const IndexType &high_)
@@ -285,20 +285,20 @@ namespace schnek
     //=========== SingleArrayInstantFortranAllocation =================
     //=================================================================
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     void SingleArrayInstantFortranAllocation<T, rank>::resize(const IndexType &low_, const IndexType &high_)
     {
         this->deleteData();
         this->newData(low_, high_);
     }
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     SingleArrayInstantFortranAllocation<T, rank>::~SingleArrayInstantFortranAllocation()
     {
         this->deleteData();
     }
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     void SingleArrayInstantFortranAllocation<T, rank>::deleteData()
     {
         if (data)
@@ -307,7 +307,7 @@ namespace schnek
         size = 0;
     }
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     void SingleArrayInstantFortranAllocation<T, rank>::newData(
         const IndexType &low_,
         const IndexType &high_)
@@ -337,19 +337,19 @@ namespace schnek
     //================== SingleArrayLazyAllocation ====================
     //=================================================================
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     SingleArrayLazyAllocation<T, rank>::SingleArrayLazyAllocation()
         : data(NULL), data_fast(NULL), size(0), bufSize(0), avgSize(0.0), avgVar(0.0), r(0.05)
     {
     }
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     SingleArrayLazyAllocation<T, rank>::~SingleArrayLazyAllocation()
     {
         this->deleteData();
     }
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     void SingleArrayLazyAllocation<T, rank>::resize(const IndexType &low_, const IndexType &high_)
     {
         int oldSize = size;
@@ -386,7 +386,7 @@ namespace schnek
         data_fast = data + p;
     }
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     void SingleArrayLazyAllocation<T, rank>::deleteData()
     {
         SCHNEK_TRACE_LOG(5, "Deleting pointer (" << (void *)data << "): size=" << size << " avgSize=" << avgSize << " avgVar=" << avgVar << " bufSize=" << bufSize);
@@ -397,7 +397,7 @@ namespace schnek
         bufSize = 0;
     }
 
-    template <typename T, int rank>
+    template <typename T, size_t rank>
     void SingleArrayLazyAllocation<T, rank>::newData(
         int newSize)
     {
